axfproject/requirements.txt
django==2.0.6
mysqlclient==1.4.2.post1
django-redis==4.10.0
django-redis-cache==2.0.0
djangorestframework==3.9.4
django-filter== 2.1.0
django-cors-headers==3.0.2
